+++ NEWS.md
@@ -8,6 +8,9 @@
 * Add a `field` argument to `$to_string()` methods to optionally omit field
   name comments from generated IDF strings (#77).
 
+* Add `Idf$reorder()` to reorder objects within a class without changing object
+  IDs (#66).
+
 ## Internal refactor
 
 * Use cli conditions for eplusr verbose logging, package warnings, errors, and